Update and rename zeltrezRates.js to zelcoreRates.js
